pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.2.0",
    "fastapi>=0.115.12",
    "flask-migrate>=4.1.0",
    "flask>=3.1.1",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "psycopg2-binary>=2.9.10",
    "sqlalchemy>=2.0.41",
    "uvicorn>=0.34.2",
    "werkzeug>=3.1.3",
    "flask-login>=0.6.3",
    "pydantic>=2.11.4",
    "python-jose>=3.4.0",
    "pyjwt>=2.10.1",
]
